particles

+++ code/datums/status_effects/status_effect.dm
@@ -28,8 +28,8 @@
 	var/atom/movable/screen/alert/status_effect/linked_alert = null
 	/// Do we self-terminate when a fullheal is called?
 	var/remove_on_fullheal = FALSE
-	// /// If remove_on_fullheal is TRUE, what flag do we need to be removed?
-	// var/heal_flag_necessary = HEAL_STATUS
+	/// If remove_on_fullheal is TRUE, what flag do we need to be removed?
+	var/heal_flag_necessary = HEAL_STATUS
 	/// Assoc list of statkey to value
 	var/list/effectedstats = list()
 
@@ -42,7 +42,6 @@
 	if(owner)
 		LAZYADD(owner.status_effects, src)
 		RegisterSignal(owner, COMSIG_LIVING_POST_FULLY_HEAL, PROC_REF(remove_effect_on_heal))
-
 	if(!owner || !on_apply())
 		qdel(src)
 		return
@@ -130,8 +129,8 @@
 	if(!remove_on_fullheal)
 		return
 
-	// if(!heal_flag_necessary || (heal_flags & heal_flag_necessary))
-	qdel(src)
+	if(!heal_flag_necessary || (heal_flags & heal_flag_necessary))
+		qdel(src)
 
 /// Remove [seconds] of duration from the status effect, qdeling / ending if we eclipse the current world time.
 /datum/status_effect/proc/remove_duration(seconds)
